All test are passing, delted, un-needed tests
