--- frontend/react-app/src/CreateTask.test.js
@@ -1,9 +1,0 @@
-import CreateTask from "./pages/CreateTask";
-import { render, screen } from "@testing-library/react";
-test("renders Task page", () => {
-    render(<CreateTask/>);
-  
-    //check if the Create Task is rendered
-    expect(screen.getByText("Task Name")).toBeInTheDocument()
-
-  });
