--- frontend/react-app/src/tests/CreateAccount.test.js
@@ -1,8 +1,0 @@
-import CreateAccount from "../pages/CreateAccount";
-import { render, screen } from "@testing-library/react";
-
-test("renders CreateAccount page", () => {
-
-    render(<CreateAccount/>);
-
-});
